ok, this is a big commit.  there was a problem with pointers which is most of the code, but I'm also working on dealing with vulgaraties in 802.11 which is a real PITA.  At this point I think I'll call it "done", but it really needs a rewrite/clean up to deal with all the 802.11 corner cases. fixes #325 - 802.11 bugs (sorta) fixes #345 - pointer bugs

+++ src/tcpr.h
@@ -125,7 +125,7 @@
 
 /**
  * IEEE 802.1Q (Virtual Local Area Network) VLAN header, static header 
- * size: 18 bytes
+ * size: 8 bytes
  */
 struct tcpr_802_1q_hdr
 {
@@ -158,7 +158,7 @@ struct tcpr_802_1x_hdr
 /*
  *  IEEE 802.2 LLC header
  *  Link Layer Control
- *  static header size: 3 bytes
+ *  static header size: 4 bytes
  */
 struct tcpr_802_2_hdr
 {
@@ -166,7 +166,7 @@ struct tcpr_802_2_hdr
     u_int8_t llc_ssap;            /* source service access point */
 #define TCPR_SAP_STP          0x42
 #define TCPR_SAP_SNAP         0xaa
-    u_int8_t llc_control;         /* control field */
+    u_int16_t llc_control;         /* control field */
 };
